add obsolete tags
